requirements.txt
requests
pandas
python-dotenv
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
discord.py
schedule
beautifulsoup4
